+++ src/events/birth/forms/printForm/collectors.ts
@@ -18,7 +18,7 @@ import {
   not
 } from '@opencrvs/toolkit/events'
 import { InformantType, InformantTypeKey } from '../pages/informant'
-import { informantMessageDescriptors } from '@countryconfig/events/common/messages'
+import { informantMessageDescriptors } from '@countryconfig/events/utils'
 
 const onlyMotherExist = (informantType: InformantTypeKey) => {
   return {
