+++ src/events/death/forms/correctionForm/requester.ts
@@ -17,7 +17,7 @@ import {
   FieldType
 } from '@opencrvs/toolkit/events'
 import { InformantType, InformantTypeKey } from '../pages/informant'
-import { informantMessageDescriptors } from '@countryconfig/events/common/messages'
+import { informantMessageDescriptors } from '@countryconfig/events/utils'
 import { IdType, idTypeOptions } from '@countryconfig/events/person'
 import { nationalIdValidator } from '@countryconfig/events/birth/validators'
