+++ src/events/death/forms/pages/informant.ts
@@ -22,7 +22,7 @@ import {
   user
 } from '@opencrvs/toolkit/events'
 import { not } from '@opencrvs/toolkit/conditionals'
-import { createSelectOptions, emptyMessage } from '../../../utils'
+
 import {
   farajalandNameConfig,
   invalidNameValidator,
@@ -36,8 +36,10 @@ import {
 } from '../../../person'
 import {
   defaultStreetAddressConfiguration,
-  getNestedFieldValidators
-} from '@countryconfig/events/street-address-configuration'
+  getNestedFieldValidators,
+  createSelectOptions,
+  emptyMessage
+} from '@countryconfig/events/utils'
 
 export const InformantType = {
   SPOUSE: 'SPOUSE',
