+++ src/events/utils/certificate/handlebars/handler.ts
